# api/openapi.yaml
openapi: 3.1.0
info:
  title: Relay Station Control API
  version: 0.2.0
  description: |
    Internal API for the single-environment Relay Station Control.

    Except for operations that explicitly set `security: []`, requests require an
    enabled `super_admin` session that satisfies the environment MFA policy. Unsafe
    methods additionally require the session-bound `X-CSRF-Token` header.
security:
  - SessionCookie: []
...
paths:
...
  /api/environment:
    get:
      operationId: getEnvironment
      summary: Read the bound Control environment
      tags: [assets]
      responses:
        "200":
          description: The database-backed environment identity and display metadata.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentAsset"
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/assets/gateway:
    get:
      operationId: getGatewayAsset
      summary: Read the single registered Gateway
      tags: [assets]
      responses:
        "200":
          description: The registered Gateway or an explicit not-registered state.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayAssetResponse"
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
...
  /api/assets/nodes:
    get:
      operationId: listNodeAssets
      summary: List registered Relay Nodes
      tags: [assets]
      parameters:
        - $ref: "#/components/parameters/AssetPageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - name: node_type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NodeType"
        - name: capability
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NodeCapability"
        - name: monitoring_active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: A stable instance-ID ordered page from one database snapshot.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeAssetListResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/assets/nodes/{instance_id}:
    get:
      operationId: getNodeAsset
      summary: Read one registered Relay Node
      tags: [assets]
      parameters:
        - $ref: "#/components/parameters/NodeInstanceId"
      responses:
        "200":
          description: The requested Relay Node.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeAsset"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/assets/drivers:
    get:
      operationId: listNodeDrivers
      summary: List registered Node Driver contracts and capabilities
      tags: [assets]
      responses:
        "200":
          description: The complete registered Driver contract catalog.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDriverListResponse"
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/assets/provider-policies/current:
    get:
      operationId: getCurrentProviderInventoryPolicy
      summary: Read the current Provider inventory policy for one Driver scope
      tags: [assets]
      parameters:
        - name: node_type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/NodeType"
        - name: driver_contract_version
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/DriverContractVersion"
      responses:
        "200":
          description: The currently effective policy or an explicit not-configured state.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentProviderInventoryPolicyResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-relay_control_session
      description: |
        Opaque administrator session cookie. In production it is set with Secure,
        HttpOnly, SameSite=Strict, Path=/, no Domain, a 30-minute idle timeout, and a
        12-hour absolute timeout. It is never a bearer token for service identities.
  parameters:
...
    PageCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 512
    AssetPageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    NodeInstanceId:
      name: instance_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
...
    NoStore:
      description: Sensitive authentication and administrator responses are never cacheable.
      required: true
      schema:
        type: string
        enum:
          - no-store
...
    RequestId:
      description: Stable request correlation identifier safe to show to an administrator.
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
...
  responses:
    Error:
      description: Request rejected using the uniform error envelope.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          description: Present for a rate-limited request, in seconds.
          schema:
            type: integer
            minimum: 1
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    EnvironmentAsset:
      type: object
      additionalProperties: false
      required: [environment_id, name, environment_type]
      properties:
        environment_id:
          type: string
          minLength: 1
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 100
        environment_type:
          type: string
          enum: [dev, staging, production]
    GatewayAssetResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [registered, not_registered]
        gateway:
          $ref: "#/components/schemas/GatewayAsset"
    GatewayAsset:
      type: object
      additionalProperties: false
      required: [instance_id, lifecycle_status, revision, display_name, management_endpoint, secret_configured, created_at, updated_at, retired_at, retired_by, retire_reason]
      properties:
        instance_id:
          type: string
          format: uuid
        lifecycle_status:
          type: string
          enum: [active, retired]
        revision:
          $ref: "#/components/schemas/AssetRevision"
        display_name:
          $ref: "#/components/schemas/DisplayName"
        management_endpoint:
          $ref: "#/components/schemas/ManagementEndpoint"
        secret_configured:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        retired_at:
          type: [string, "null"]
          format: date-time
        retired_by:
          type: [string, "null"]
          format: uuid
        retire_reason:
          type: [string, "null"]
          enum: [administrator_retire, replacement, null]
    AssetRevision:
      type: string
      pattern: "^[1-9][0-9]{0,18}$"
      maxLength: 19
...
    NodeAssetListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NodeAsset"
        next_cursor:
          type: [string, "null"]
          maxLength: 512
    NodeAsset:
      type: object
      additionalProperties: false
      required:
        - instance_id
        - display_name
        - node_type
        - driver_contract_version
        - management_endpoint
        - secret_configured
        - capabilities
        - monitoring
        - created_at
        - updated_at
      properties:
        instance_id:
          type: string
          format: uuid
        display_name:
          $ref: "#/components/schemas/DisplayName"
        node_type:
          $ref: "#/components/schemas/NodeType"
        driver_contract_version:
          $ref: "#/components/schemas/DriverContractVersion"
        management_endpoint:
          $ref: "#/components/schemas/ManagementEndpoint"
        secret_configured:
          type: boolean
        capabilities:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/NodeCapability"
        monitoring:
          $ref: "#/components/schemas/NodeMonitoringStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NodeMonitoringStatus:
      type: object
      additionalProperties: false
      required: [active]
      properties:
        active:
          type: boolean
        effective_from:
          type: [string, "null"]
          format: date-time
        effective_to:
          type: [string, "null"]
          format: date-time
    NodeDriverListResponse:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NodeDriver"
    NodeDriver:
      type: object
      additionalProperties: false
      required: [node_type, driver_contract_version, display_name, lifecycle_status, capabilities, created_at]
      properties:
        node_type:
          $ref: "#/components/schemas/NodeType"
        driver_contract_version:
          $ref: "#/components/schemas/DriverContractVersion"
        display_name:
          $ref: "#/components/schemas/DisplayName"
        lifecycle_status:
          type: string
          enum: [active, deprecated, retired]
        capabilities:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/NodeCapability"
        created_at:
          type: string
          format: date-time
    CurrentProviderInventoryPolicyResponse:
      type: object
      additionalProperties: false
      required: [status, node_type, driver_contract_version]
      properties:
        status:
          type: string
          enum: [configured, not_configured]
        node_type:
          $ref: "#/components/schemas/NodeType"
        driver_contract_version:
          $ref: "#/components/schemas/DriverContractVersion"
        policy:
          $ref: "#/components/schemas/ProviderInventoryPolicy"
    ProviderInventoryPolicy:
      type: object
      additionalProperties: false
      required: [version_id, active_providers, out_of_scope_providers, effective_from, created_at]
      properties:
        version_id:
          type: string
          format: uuid
        active_providers:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ProviderName"
        out_of_scope_providers:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ProviderName"
        effective_from:
          type: string
          format: date-time
        effective_to:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
    NodeType:
      type: string
      minLength: 2
      maxLength: 64
      pattern: "^[a-z0-9][a-z0-9._-]*$"
    DriverContractVersion:
      type: string
      minLength: 1
      maxLength: 64
      pattern: "^[a-z0-9][a-z0-9._-]*$"
    NodeCapability:
      type: string
      enum: [management_health_read, management_account_inventory_read]
    ProviderName:
      type: string
      minLength: 1
      maxLength: 64
      pattern: "^[a-z0-9][a-z0-9._-]*$"
    ManagementEndpoint:
      type: string
      format: uri
      minLength: 8
      maxLength: 2048
      pattern: "^https?://[^/?#]+(?:/[^?#]*)?$"
...
    DisplayName:
      type: string
      minLength: 1
      maxLength: 100
      pattern: "^\\S(?:.*\\S)?$|^\\S$"
...
    ErrorCode:
      type: string
      enum:
        - authentication_failed
        - mfa_required
        - challenge_expired
        - rate_limited
        - unauthorized
        - forbidden
        - csrf_invalid
        - validation_failed
        - conflict
        - bootstrap_unavailable
        - reauthentication_required
        - last_administrator_protected
        - administrator_self_disable_forbidden
        - not_found
        - internal_error
        - temporarily_unavailable
        - directory_unavailable
        - directory_stale
        - account_not_found
        - node_conflict
        - account_conflict
        - already_unbound
        - no_current_binding
        - asset_not_found
        - asset_retired
        - current_gateway_exists
        - duplicate_identity
        - stale_revision
        - revision_exhausted
        - command_conflict
        - cursor_stale
        - invalid_endpoint
        - secret_configuration_invalid
        - probe_timeout
        - probe_failed
        - service_unavailable
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          minLength: 1
          maxLength: 256
        request_id:
          type: string
          minLength: 8
          maxLength: 128
          pattern: "^[A-Za-z0-9._:-]+$"
        retry_after_seconds:
          type: integer
          minimum: 1
          maximum: 86400
...